fixed / moved components into different folders. Also removed old Components
