+++ react_app/src/Components/Cards/Cards.test.js
@@ -1,0 +1,9 @@
+import { render, screen } from '@testing-library/react';
+import Cards from './Cards';
+
+test('making sure utcConverter returns a number 0-7', () => { 
+    render(<Cards/>)
+    
+    
+
+})
