--- react_app/src/Components/Navbar/MenuItems.js
@@ -1,17 +1,0 @@
-export const MenuItems = [
-    {
-        title: 'Home',
-        url: '#',
-        cName: 'nav-links'
-    },
-    {
-        title: 'Page1',
-        url: '#',
-        cName: 'nav-links'
-    },
-    {
-        title: 'Page2',
-        url: '#',
-        cName: 'nav-links'
-    },
-]
