--- react_app/src/Components/Navbar/Navbar.css
@@ -1,47 +1,0 @@
-.NavbarItems{
-    background: linear-gradient(90deg, rgb(110,96,253), rgba(73,63,230,1) 100%);
-    height: 180px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 1.2rem;
-}
-.navbar-logo{
-    color: #fff;
-    justify-self: start;
-    margin-left: 20px;
-    cursor: pointer;
-
-}
-.fa-react{
-    margin-left: 0.5rem;
-    font-size: 1.6rem;
-}
-.nav-menu{
-    display: grid;
-    grid-template-columns: repeat(5, auto);
-    grid-gap: 10px;
-    list-style: none;
-    text-align: center;
-    width: 70vw;
-    justify-content: end;
-    margin-right: 2rem;
-
-}
-.nav-links{
-    color:white;
-    text-decoration: none;
-    padding: 0.5rem 1rem;
-}
-.nav-links:hover{
-    background-color: blueviolet;
-    border-radius: 4px;
-    transition: all 0.2s ease-out;
-
-}
-.fa-bars{
-    color: black;
-}
-.menu-icon{
-    display: none;
-}
