--- react_app/src/Components/Navbar/Navbar.js
@@ -1,34 +1,0 @@
-import React, { Component } from 'react';
-import './Navbar.css';
-import {MenuItems} from "./MenuItems";
-
-class Navbar extends Component{
-    state = { clicked: false }
-
-    handleClick = () =>{
-        this.setState({clicked: !this.state.clicked })
-    }
-    render(){
-        return(
-            <nav className="NavbarItems">
-                <h1 className="navbar-logo">Weather App<i className="fab fa-react"></i></h1>
-                <div className="menu-icon" onClick={this.handleClick}>
-                    <i className={ this.state.clicked ? 'fas fa-times' : 'fas fa-bars'}></i>
-                </div>
-                <ul className={this.state.clicked ? 'nav-menu active' : 'nav-menu'}>
-                    {MenuItems.map((item, index) => {
-                        return(
-                            <li>
-                                <a className={item.cName} href= {item.url}>
-                                    {item.title}
-                                </a>
-                            </li>
-                        )                       
-                    })}
-                </ul>
-            </nav>
-        )
-    }
-}
-
-export default Navbar
